+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Switch from setup.cfg to pyproject.toml 
 
+### Fixed
+
+- Fix frontend translation of Enroll now for external LMS backend
+
 ## [2.25.0-beta.1]
 
 ### Added